py_code/requirements.txt
absl-py==1.0.0
appnope @ file:///opt/concourse/worker/volumes/live/0291c9e1-4b15-459f-623e-2770f55be269/volume/appnope_1594338395037/work
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cachetools==5.0.0
certifi==2020.6.20
charset-normalizer==2.0.12
cma==3.2.1
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.11.2
importlib-resources==5.4.0
ipykernel @ file:///opt/concourse/worker/volumes/live/88f541d3-5a27-498f-7391-f2e50ca36560/volume/ipykernel_1596206680118/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/90527372-a5f7-4d48-68c2-a99de2ba31c7/volume/ipython_1599056199422/work
ipython-genutils==0.2.0
jedi @ file:///opt/concourse/worker/volumes/live/fb2e45f1-1dff-42fa-6cab-bd094a4c4372/volume/jedi_1644315264796/work
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
mistune==0.8.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pexpect @ file:///opt/concourse/worker/volumes/live/8701bb20-ad87-46c7-5108-30c178cf97e5/volume/pexpect_1594383388344/work
pickleshare @ file:///opt/concourse/worker/volumes/live/93ec39d8-05bb-4f84-7efc-98735bc39b70/volume/pickleshare_1594384101884/work
Pillow==9.0.1
pipreqs==0.4.11
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.19.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2021.3
PyYAML==6.0
pyzmq==19.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
tornado==6.0.4
tqdm==4.63.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
tweepy==4.8.0
typing-extensions==4.1.1
urllib3==1.26.8
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.13.3
yarg==0.1.9
zipp==3.7.0
